Completed T10 - Added GET users. Refactored comments endpoint into correct model/controllers.
