+++ controller/articles.controller.js
@@ -4,7 +4,6 @@ const {
   fetchArticleComments,
   publishArticleComment,
   modelPatchArticleVotes,
-  modelDeleteComment,
 } = require("../model/articles.model");
 
 exports.getArticle = (request, response, next) => {
@@ -62,17 +61,6 @@ exports.patchArticleVotes = (request, response, next) => {
   modelPatchArticleVotes(patchNum, patchBody)
     .then((result) => {
       response.status(200).send({ article: result });
-    })
-    .catch((err) => {
-      next(err);
-    });
-};
-
-exports.deleteComment = (request, response, next) => {
-  const commentNum = request.params.comment_id;
-  modelDeleteComment(commentNum)
-    .then(() => {
-      response.status(204).send({ message: "Deletion Successful!" });
     })
     .catch((err) => {
       next(err);
