+++ endpoints.json
@@ -91,5 +91,13 @@
     "exampleResponse": {
       "message": "Deletion Successful!"
     }
+  },
+  "GET /api/users": {
+    "description": "Returns an array of all users",
+    "exampleResponse": {
+      "username": "rogersop",
+      "name": "paul",
+      "avatar_url": "https://avatars2.githubusercontent.com/u/24394918?s=400&v=4"
+    }
   }
 }
