+++ model/articles.model.js
@@ -70,18 +70,6 @@ function modelPatchArticleVotes(patchNum, patchBody) {
     });
 }
 
-function modelDeleteComment(commentNum) {
-  return db
-    .query("DELETE FROM comments WHERE comment_id = $1 RETURNING *", [
-      commentNum,
-    ])
-    .then((result) => {
-      if (result.rowCount === 0) {
-        return Promise.reject({ status: 404, message: "Comment Not Found" });
-      }
-    });
-}
-
 module.exports = {
   fetchArticle,
   fetchAllArticles,
@@ -89,5 +77,4 @@ module.exports = {
   publishArticleComment,
   modelPatchArticleVotes,
   patchArticleVotes,
-  modelDeleteComment,
 };
